+++ Sources/xcodeproj/Objects/BuildPhase/PBXBuildFile.swift
@@ -10,9 +10,7 @@ public final class PBXBuildFile: PBXObject {
     /// Returns the file the build file refers to.
     public var file: PBXFileElement? {
         get {
-            return fileReference.flatMap { (reference) -> PBXFileElement? in
-                try? reference.object()
-            }
+            return fileReference?.getObject()
         }
         set {
             fileReference = newValue?.reference
@@ -63,8 +61,8 @@ extension PBXBuildFile {
     /// - Returns: file name.
     /// - Throws: an error if the name cannot be obtained.
     func fileName() throws -> String? {
-        guard let fileElement: PBXFileElement? = try? fileReference?.object() else { return nil }
-        return fileElement?.fileName()
+        guard let fileElement: PBXFileElement = fileReference?.getObject() else { return nil }
+        return fileElement.fileName()
     }
 
     /// Returns the type of the build phase the build file belongs to.
@@ -118,7 +116,7 @@ extension PBXBuildFile: PlistSerializable {
         var dictionary: [CommentedString: PlistValue] = [:]
         dictionary["isa"] = .string(CommentedString(PBXBuildFile.isa))
         if let fileReference = fileReference {
-            let fileElement: PBXFileElement? = try? fileReference.object()
+            let fileElement: PBXFileElement? = fileReference.getObject()
             dictionary["fileRef"] = .string(CommentedString(fileReference.value, comment: fileElement?.fileName()))
         }
         if let settings = settings {