+++ Sources/xcodeproj/Objects/Project/PBXProj.swift
@@ -25,9 +25,7 @@ public final class PBXProj: Decodable {
             rootObjectReference = newValue?.reference
         }
         get {
-            return rootObjectReference.flatMap { (reference) -> PBXProject? in
-                try? reference.object()
-            }
+            return rootObjectReference?.getObject()
         }
     }
 
@@ -119,13 +117,13 @@ public extension PBXProj {
 
     /// Returns root project.
     public func rootProject() throws -> PBXProject? {
-        return try rootObjectReference?.object()
+        return try rootObjectReference?.getThrowingObject()
     }
 
     /// Returns root project's root group.
     public func rootGroup() throws -> PBXGroup? {
         let project = try rootProject()
-        return try project?.mainGroupReference.object()
+        return try project?.mainGroupReference.getThrowingObject()
     }
 
     /// Adds a new object to the project.