+++ Sources/xcodeproj/Objects/Project/PBXProject.swift
@@ -15,7 +15,7 @@ public final class PBXProject: PBXObject {
             buildConfigurationListReference = newValue.reference
         }
         get {
-            return try? buildConfigurationListReference.object()
+            return buildConfigurationListReference.getObject()
         }
     }
 
@@ -40,7 +40,7 @@ public final class PBXProject: PBXObject {
             mainGroupReference = newValue.reference
         }
         get {
-            return try? mainGroupReference.object()
+            return mainGroupReference.getObject()
         }
     }
 
@@ -53,9 +53,7 @@ public final class PBXProject: PBXObject {
             productsGroupReference = newValue?.reference
         }
         get {
-            return productsGroupReference.flatMap { (reference) -> PBXGroup? in
-                try? reference.object()
-            }
+            return productsGroupReference?.getObject()
         }
     }
 
@@ -78,7 +76,7 @@ public final class PBXProject: PBXObject {
         }
         get {
             return projectReferences.map { project in
-                project.flatMapValues({ try? $0.object() })
+                project.flatMapValues({ $0.getObject() })
             }
         }
     }
@@ -94,10 +92,10 @@ public final class PBXProject: PBXObject {
     /// Project targets.
     public var targets: [PBXTarget] {
         set {
-            targetReferences = newValue.map({ $0.reference })
+            targetReferences = newValue.references()
         }
         get {
-            return targetReferences.compactMap({ try? $0.object() })
+            return targetReferences.objects()
         }
     }
 
@@ -118,7 +116,7 @@ public final class PBXProject: PBXObject {
         } get {
             var attributes: [PBXTarget: [String: Any]] = [:]
             targetAttributeReferences.forEach {
-                if let object: PBXTarget = try? $0.key.object() {
+                if let object: PBXTarget = $0.key.getObject() {
                     attributes[object] = $0.value
                 }
             }
@@ -197,7 +195,7 @@ public final class PBXProject: PBXObject {
         self.projectDirPath = projectDirPath
         projectReferences = projects.map({ project in project.mapValues({ $0.reference }) })
         self.projectRoots = projectRoots
-        targetReferences = targets.map({ $0.reference })
+        targetReferences = targets.references()
         self.attributes = attributes
         targetAttributeReferences = [:]
         super.init()
@@ -290,10 +288,10 @@ extension PBXProject: PlistSerializable {
             dictionary["knownRegions"] = PlistValue.array(knownRegions
                 .map { .string(CommentedString("\($0)")) })
         }
-        let mainGroupObject: PBXGroup? = try? mainGroupReference.object()
+        let mainGroupObject: PBXGroup? = mainGroupReference.getObject()
         dictionary["mainGroup"] = .string(CommentedString(mainGroupReference.value, comment: mainGroupObject?.fileName()))
         if let productsGroupReference = productsGroupReference {
-            let productRefGroupObject: PBXGroup? = try? productsGroupReference.object()
+            let productRefGroupObject: PBXGroup? = productsGroupReference.getObject()
             dictionary["productRefGroup"] = .string(CommentedString(productsGroupReference.value,
                                                                     comment: productRefGroupObject?.fileName()))
         }
@@ -308,7 +306,7 @@ extension PBXProject: PlistSerializable {
         }
         dictionary["targets"] = PlistValue.array(targetReferences
             .map { targetReference in
-                let target: PBXTarget? = try? targetReference.object()
+                let target: PBXTarget? = targetReference.getObject()
                 return .string(CommentedString(targetReference.value, comment: target?.name))
         })
 
@@ -334,9 +332,9 @@ extension PBXProject: PlistSerializable {
             guard let productGroupReference = reference["ProductGroup"], let projectRef = reference["ProjectRef"] else {
                 return nil
             }
-            let producGroup: PBXGroup? = try? productGroupReference.object()
+            let producGroup: PBXGroup? = productGroupReference.getObject()
             let groupName = producGroup?.fileName()
-            let project: PBXFileElement? = try? projectRef.object()
+            let project: PBXFileElement? = projectRef.getObject()
             let fileRefName = project?.fileName()
 
             return [